// server/src/modules/profile/profile.enum.ts
export enum IncomeAverage {
  OPTION_ONE = '-279718',
  OPTION_TWO = '280000-1400000',
  OPTION_THREE = '+14000000',
}

export enum ExpensesAverage {
  OPTION_ONE = '-30%',
  OPTION_TWO = '30%-60%',
  OPTION_THREE = '+60%',
}

// export enum MonthlyContribution {
//   OPTION_ONE = '-10%',
//   OPTION_TWO = '10%-20%',
//   OPTION_THREE = '+20%',
// }

export enum TargetPeriod {
  CORTO = 'Menos de 1 año',
  MEDIANO = 'De 1 a 3 años',
  LARGO = 'Más 3 de años',
}